output memory usage of matrix version in DEBUG_MEMUSAGE put back (size_t) cast in INDEX2D macro needed for very large matrices

+++ src/utils.h
@@ -75,7 +75,7 @@ extern "C" {
 #define MAX(a, b) ( (a) > (b) ? (a) : (b) )
   
 /* Index into 2d n x n array stored row-major in contiguous memory  */
-#define INDEX2D(i,j,n) ( ((i)*(n) + (j)) )
+#define INDEX2D(i,j,n) ( ((size_t)(i)*(n) + (j)) )
 
   
 /*****************************************************************************
